답변 api 구현 및 도메인 변경에 따른 코드 변경

See merge request s12-webmobile2-sub1/S12P11A609!34

+++ backend/src/main/java/com/ssafy/d3v/backend/like/dto/LikesRequest.java
@@ -1,4 +1,4 @@
 package com.ssafy.d3v.backend.like.dto;
 
-public record LikesRequest(int memberId) {
+public record LikesRequest(long memberId) {
 }
