+++ backend/src/main/java/com/ssafy/d3v/backend/like/service/LikesService.java
@@ -3,6 +3,6 @@ package com.ssafy.d3v.backend.like.service;
 import com.ssafy.d3v.backend.like.dto.LikesRequest;
 
 public interface LikesService {
-    void create(int answerId, LikesRequest request);
-    void delete(int answerId);
+    void create(long answerId, LikesRequest request);
+    void delete(long answerId);
 }
